+++ Deeploy/Targets/Generic/Layers.py
@@ -69,15 +69,16 @@ class GELULayer(ONNXLayer):
         super().__init__(maps)
 
     def computeOps(self):
-        compAbs = self.mapper.parser.operatorRepresentation['size']
-        compAdd = self.mapper.parser.operatorRepresentation['size']
-        compSqr = self.mapper.parser.operatorRepresentation['size']
-        compMul = self.mapper.parser.operatorRepresentation['size']
-        compAdd = self.mapper.parser.operatorRepresentation['size']
-        compMul2 = self.mapper.parser.operatorRepresentation['size']
-        compAdd2 = self.mapper.parser.operatorRepresentation['size']
-        compDiv = self.mapper.parser.operatorRepresentation['size']
-        return compAbs + compAdd + compSqr + compMul + compAdd + compMul2 + compAdd2 + compDiv
+        size = self.mapper.parser.operatorRepresentation['size']
+        # RW: Sigmoid approximation
+        mul1 = size  # Multiply by 1.702
+        neg = size  # Negate the result
+        exp = size  # Compute exponential
+        add = size  # Add 1
+        div = size  # Division for sigmoid
+        mul2 = size  # Final multiplication by x
+
+        return mul1 + neg + exp + add + div + mul2
 
 
 class iHardswishLayer(ONNXLayer):
@@ -120,11 +121,38 @@ class SoftmaxLayer(ONNXLayer):
     def __init__(self, maps: List[NodeMapper]):
         super().__init__(maps)
 
+    def computeOps(self):
+
+        size = self.mapper.parser.operatorRepresentation['size']
+        last_dim_length = self.mapper.parser.operatorRepresentation['lastDimLength']
+        batch_size = size // last_dim_length
+
+        max_ops = last_dim_length - 1
+        exp_ops = last_dim_length * 2
+        sum_ops = last_dim_length - 1
+        div_ops = last_dim_length
+        ops_per_batch = max_ops + exp_ops + sum_ops + div_ops
+        total_ops = ops_per_batch * batch_size
+
+        return total_ops
+
 
 class SoftmaxGradLayer(ONNXLayer):
 
     def __init__(self, maps: List[NodeMapper]):
         super().__init__(maps)
+
+    def computeOps(self):
+        input_size = self.mapper.parser.operatorRepresentation['size']
+
+        # SoftmaxGrad operation: dy * (y - (y * sum(dy * y)))
+        mul_ops = input_size
+        sum_ops = input_size
+        broadcast_mul_ops = input_size
+        sub_ops = input_size
+        final_mul_ops = input_size
+
+        return mul_ops + sum_ops + broadcast_mul_ops + sub_ops + final_mul_ops
 
 
 class ITAMaxLayer(ONNXLayer):
@@ -317,6 +345,9 @@ class MulLayer(ONNXLayer):
             inputShapes[0] = inputShapes[1]
         return (inputShapes, outputShapes)
 
+    def computeOps(self):
+        return self.mapper.parser.operatorRepresentation['size']
+
 
 class ConvLayer(ONNXLayer):
 
@@ -374,6 +405,14 @@ class MaxPoolLayer(ONNXLayer):
     def __init__(self, maps: List[NodeMapper]):
         super().__init__(maps)
 
+    def computeOps(self):
+        kernel_shape = self.mapper.parser.operatorRepresentation['kernel_shape']
+        elements_per_window = int(np.prod(kernel_shape))
+        data_out_size = self.mapper.parser.operatorRepresentation['data_out_size']
+        comparisons_per_window = elements_per_window - 1
+        total_ops = data_out_size * comparisons_per_window
+        return total_ops
+
 
 class ReduceMeanLayer(ONNXLayer):
 
@@ -402,6 +441,9 @@ class ReluLayer(ONNXLayer):
 
     def __init__(self, maps: List[NodeMapper]):
         super().__init__(maps)
+
+    def computeOps(self):
+        return self.mapper.parser.operatorRepresentation['size']
 
 
 class LayerNormLayer(ONNXLayer):